# docs/sprint-1/openapi.yaml
openapi: 3.1.0
info:
  title: SmartSweep API
  version: 0.1.0
paths:
  /api/v1/auth/register:
    post:
      tags:
      - Auth
      summary: Register a new citizen account
      description: 'Register a new citizen account.


        Public self-registration is restricted to Citizens only.

        Crew members are onboarded and provisioned directly by Administrators

        via the operational management portal.'
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - Auth
      summary: Authenticate user and issue JWT access and refresh tokens
      description: Authenticate user with email and password.
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh access token using valid refresh token
      description: Issue new token pair using a valid refresh token.
      operationId: refresh_token_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - Auth
      summary: Get profile of current logged-in user
      description: Return currently authenticated user details.
      operationId: get_me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
      security:
      - HTTPBearer: []
  /api/v1/auth/admin-only:
    get:
      tags:
      - Auth
      summary: Admin-restricted endpoint for role authorization checks
      description: Endpoint accessible only by Admin role.
      operationId: admin_only_check_api_v1_auth_admin_only_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Admin Only Check Api V1 Auth Admin Only Get
      security:
      - HTTPBearer: []
  /api/v1/auth/crew-only:
    get:
      tags:
      - Auth
      summary: Crew and Admin restricted endpoint
      description: Endpoint accessible by Crew and Admin roles.
      operationId: crew_only_check_api_v1_auth_crew_only_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Crew Only Check Api V1 Auth Crew Only Get
      security:
      - HTTPBearer: []
...
  /api/v1/bulk-pickups:
    post:
      tags:
      - Bulk Pickups
      summary: Create Bulk Pickup
      operationId: create_bulk_pickup_api_v1_bulk_pickups_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkPickupCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkPickupRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Bulk Pickups
      summary: List Bulk Pickups
      operationId: list_bulk_pickups_api_v1_bulk_pickups_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status_filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Filter
      - name: ward_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_BulkPickupRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/bulk-pickups/{pickup_id}:
    get:
      tags:
      - Bulk Pickups
      summary: Get Bulk Pickup
      operationId: get_bulk_pickup_api_v1_bulk_pickups__pickup_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: pickup_id
        in: path
        required: true
        schema:
          type: integer
          title: Pickup Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkPickupRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Bulk Pickups
      summary: Update Bulk Pickup
      operationId: update_bulk_pickup_api_v1_bulk_pickups__pickup_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: pickup_id
        in: path
        required: true
        schema:
          type: integer
          title: Pickup Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkPickupUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkPickupRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/bulk-pickups/{pickup_id}/cancel:
    post:
      tags:
      - Bulk Pickups
      summary: Cancel Bulk Pickup
      operationId: cancel_bulk_pickup_api_v1_bulk_pickups__pickup_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: pickup_id
        in: path
        required: true
        schema:
          type: integer
          title: Pickup Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkPickupRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/bulk-pickups/{pickup_id}/assign:
    post:
      tags:
      - Bulk Pickups
      summary: Assign Bulk Pickup
      operationId: assign_bulk_pickup_api_v1_bulk_pickups__pickup_id__assign_post
      security:
      - HTTPBearer: []
      parameters:
      - name: pickup_id
        in: path
        required: true
        schema:
          type: integer
          title: Pickup Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkPickupAssign'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkPickupRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BulkPickupAssign:
      properties:
        worker_id:
          type: integer
          title: Worker Id
        vehicle_id:
          type: integer
          title: Vehicle Id
      type: object
      required:
      - worker_id
      - vehicle_id
      title: BulkPickupAssign
      description: Payload for dispatching a crew member and vehicle to a pickup.
    BulkPickupCategory:
      type: string
      enum:
      - general
      - e_waste
      - construction_debris
      - scrap_metal
      title: BulkPickupCategory
      description: Waste category. Mirrors the ORM enum.
    BulkPickupCreate:
      properties:
        category:
          $ref: '#/components/schemas/BulkPickupCategory'
          default: general
        load_band:
          $ref: '#/components/schemas/BulkPickupLoadBand'
        address:
          type: string
          maxLength: 255
          minLength: 1
          title: Address
        ward_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        preferred_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Preferred Date
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - load_band
      - address
      title: BulkPickupCreate
      description: Payload for a citizen requesting a bulk pickup.
    BulkPickupLoadBand:
      type: string
      enum:
      - small
      - medium
      - large
      - extra_large
      title: BulkPickupLoadBand
      description: Load size. Mirrors the ORM enum.
    BulkPickupRead:
      properties:
        id:
          type: integer
          title: Id
        requested_by_user_id:
          type: integer
          title: Requested By User Id
        ward_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
        category:
          type: string
          title: Category
        load_band:
          type: string
          title: Load Band
        address:
          type: string
          title: Address
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        preferred_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Preferred Date
        status:
          type: string
          title: Status
        assigned_vehicle_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assigned Vehicle Id
        assigned_worker_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assigned Worker Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        scheduled_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Scheduled At
        collected_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Collected At
        cancelled_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cancelled At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - requested_by_user_id
      - category
      - load_band
      - address
      - status
      - created_at
      - updated_at
      title: BulkPickupRead
      description: Bulk pickup response schema.
    BulkPickupStatus:
      type: string
      enum:
      - requested
      - scheduled
      - collected
      - cancelled
      title: BulkPickupStatus
      description: Lifecycle states exposed through the API. Mirrors the ORM enum.
    BulkPickupUpdate:
      properties:
        status:
          anyOf:
          - $ref: '#/components/schemas/BulkPickupStatus'
          - type: 'null'
        assigned_vehicle_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assigned Vehicle Id
        assigned_worker_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assigned Worker Id
        scheduled_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Scheduled At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      title: BulkPickupUpdate
      description: Payload for crew/admin managing a pickup's lifecycle.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          minLength: 1
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
      description: Payload for user login.
...
    PageMeta:
      properties:
        page:
          type: integer
          minimum: 1.0
          title: Page
          description: 1-based index of the returned page.
        page_size:
          type: integer
          minimum: 1.0
          title: Page Size
          description: Maximum items per page.
        total:
          type: integer
          minimum: 0.0
          title: Total
          description: Total matching rows, ignoring pagination.
        total_pages:
          type: integer
          minimum: 0.0
          title: Total Pages
          description: ceil(total / page_size).
      type: object
      required:
      - page
      - page_size
      - total
      - total_pages
      title: PageMeta
      description: Pagination counters. ``total_pages`` is derived, never supplied.
    Page_BulkPickupRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/BulkPickupRead'
          type: array
          title: Items
        meta:
          $ref: '#/components/schemas/PageMeta'
      type: object
      required:
      - items
      - meta
      title: Page[BulkPickupRead]
...
    RefreshTokenRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshTokenRequest
      description: Payload for refreshing an access token.
...
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      - refresh_token
      title: Token
      description: Token response schema containing access and refresh tokens.
...
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        full_name:
          type: string
          maxLength: 255
          minLength: 1
          title: Full Name
        role:
          $ref: '#/components/schemas/UserRole'
          default: citizen
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        ward_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
        password:
          type: string
          maxLength: 128
          minLength: 6
          title: Password
      type: object
      required:
      - email
      - full_name
      - password
      title: UserCreate
      description: Payload for user registration or creation.
    UserRead:
      properties:
        email:
          type: string
          format: email
          title: Email
        full_name:
          type: string
          maxLength: 255
          minLength: 1
          title: Full Name
        role:
          $ref: '#/components/schemas/UserRole'
          default: citizen
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        ward_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
        id:
          type: integer
          title: Id
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - email
      - full_name
      - id
      - is_active
      - created_at
      - updated_at
      title: UserRead
      description: User response schema.
    UserRole:
      type: string
      enum:
      - citizen
      - crew
      - admin
      title: UserRole
      description: User roles for Role-Based Access Control (RBAC).
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer